add company to graphql

+++ src/main/java/hello/entity/Company.java
@@ -12,13 +12,13 @@ import java.util.Date;
  **/
 @Data
 public class Company {
-  private Integer id;
+  private Long id;
   private String name;
   private String shortName;
   private String address;
   private String linkName;
   private String phone;
-  private Integer cityId;
+  private Long cityId;
   private Integer status;
   private Date createdAt;
   private Date updatedAt;
